requirements.txt
pyteal
py-algorand-sdk
mypy
pytest
black
django
djoser
djangorestframework # https://github.com/sunscrapers/djoser/issues/541
markdown
django-filter
django_cors_headers
djangorestframework-simplejwt